Fixed up C++ build side, now to fix FFI
