+++ jingle_sleigh/src/ffi/cpp/context.cpp
@@ -3,50 +3,12 @@
 
 #include <memory>
 #include <utility>
+#include "image_context.h"
 #include "jingle_sleigh/src/ffi/instruction.rs.h"
 #include "sleigh/loadimage.hh"
 
 
-DummyLoadImage::DummyLoadImage() : ghidra::LoadImage("jingle") {
-    img = Image{};
-}
-
-DummyLoadImage::DummyLoadImage(Image image) : ghidra::LoadImage("jingle") {
-    img = std::move(image);
-}
-
-void DummyLoadImage::loadFill(ghidra::uint1 *ptr, ghidra::int4 size, const ghidra::Address &addr) {
-    size_t offset = addr.getOffset();
-    size_t bytes_written = 0;
-    for (const auto &section: img.sections) {
-        size_t start = section.base_address;
-        size_t end = start + section.data.size();
-        if (start <= offset && offset < end) {
-            size_t len = std::min((size_t) size, (size_t) end - (size_t) offset);
-            size_t start_idx = offset - start;
-            std::memcpy(ptr, &section.data[start_idx], len);
-            offset = offset + len;
-            bytes_written += len;
-        }
-    }
-    for (size_t i = offset; i < size; ++i) {
-        ptr[i] = 0;
-    }
-    if (bytes_written == 0) {
-        ghidra::ostringstream errmsg;
-        errmsg << "Unable to load " << std::dec << size << " bytes at " << addr.getShortcut();
-        addr.printRaw(errmsg);
-        throw ghidra::DataUnavailError(errmsg.str());
-    }
-}
-
-void DummyLoadImage::adjustVma(long adjust) {}
-
-std::string DummyLoadImage::getArchType() const {
-    return "placeholder";
-}
-
-ContextFFI::ContextFFI(rust::Str slaPath) {
+ContextFFI::ContextFFI(rust::Str slaPath): sleigh(&image, &c_db) {
     ghidra::AttributeId::initialize();
     ghidra::ElementId::initialize();
 
@@ -61,8 +23,8 @@ ContextFFI::ContextFFI(rust::Str slaPath) {
     ghidra::Document *doc = documentStorage.parseDocument(sleighfilename);
     ghidra::Element *root = doc->getRoot();
     documentStorage.registerTag(root);
-    sleigh = ghidra::Sleigh(&img, &c_db);
-    sleigh.initialize(documentStorage);
+    this->sleigh = ghidra::Sleigh(&img, &c_db);
+    this->sleigh.initialize(documentStorage);
 
 }
 
@@ -94,8 +56,8 @@ rust::Str ContextFFI::getRegisterName(VarnodeInfoFFI vn) const {
     return {name};
 }
 
-std::unique_ptr<ContextFFI> makeContext(rust::Str slaPath, Image img) {
-    return std::make_unique<ContextFFI>(slaPath, std::move(img));
+std::unique_ptr<ContextFFI> makeContext(rust::Str slaPath) {
+    return std::make_unique<ContextFFI>(slaPath);
 }
 
 VarnodeInfoFFI varnodeToFFI(ghidra::VarnodeData vn) {
@@ -124,5 +86,5 @@ rust::Vec<RegisterInfoFFI> ContextFFI::getRegisters() const {
 }
 
 std::unique_ptr<SleighImage> ContextFFI::makeImageContext(Image img) {
-    return std::unique_ptr<SleighImage>(sleigh, img);
+    return std::unique_ptr<SleighImage>(new SleighImage(img, sleigh));
 }
