+++ jingle_sleigh/src/ffi/cpp/image_context.cpp
@@ -2,7 +2,7 @@
 // Created by mark denhoed on 10/10/24.
 //
 #include "image_context.h"
-#include "context.h"
+#include "dummy_load_image.h"
 
 #include <utility>
 
@@ -53,24 +53,22 @@ public:
     }
 };
 
-SleighImage::SleighImage(Image img, ghidra::Sleigh sl) {
-    sl = sl;
-    image = DummyLoadImage(img);
+SleighImage::SleighImage(Image img, ghidra::Sleigh sl): sl(sl), image(DummyLoadImage(img)) {
     sl.reset(&image, &c_db);
 }
 
 
 
-std::shared_ptr<AddrSpaceHandle> ContextFFI::getSpaceByIndex(ghidra::int4 idx) const {
-    return std::make_shared<AddrSpaceHandle>(sleigh.getSpace(idx));
+std::shared_ptr<AddrSpaceHandle> SleighImage::getSpaceByIndex(ghidra::int4 idx) const {
+    return std::make_shared<AddrSpaceHandle>(sl.getSpace(idx));
 }
 
-ghidra::int4 ContextFFI::getNumSpaces() const {
-    return sleigh.numSpaces();
+ghidra::int4 SleighImage::getNumSpaces() const {
+    return sl.numSpaces();
 }
 
-VarnodeInfoFFI ContextFFI::getRegister(rust::Str name) const {
-    ghidra::VarnodeData vn = sleigh.getRegister(name.operator std::string());
+VarnodeInfoFFI SleighImage::getRegister(rust::Str name) const {
+    ghidra::VarnodeData vn = sl.getRegister(name.operator std::string());
     VarnodeInfoFFI info;
     info.space = std::make_unique<AddrSpaceHandle>(vn.space);
     info.size = vn.size;
@@ -78,11 +76,7 @@ VarnodeInfoFFI ContextFFI::getRegister(rust::Str name) const {
     return info;
 };
 
-rust::Str ContextFFI::getRegisterName(VarnodeInfoFFI vn) const {
-    std::string name = sleigh.getRegisterName(vn.space->getRaw(), vn.offset, vn.size);
+rust::Str SleighImage::getRegisterName(VarnodeInfoFFI vn) const {
+    std::string name = sl.getRegisterName(vn.space->getRaw(), vn.offset, vn.size);
     return {name};
-}
-
-std::unique_ptr<ContextFFI> makeContext(rust::Str slaPath, Image img) {
-    return std::make_unique<ContextFFI>(slaPath, std::move(img));
 }
